Bug #48030 UDL: Remove wrong select_server() YAML
when connecting the LDAP server fails as that will just reconnect any
LDAP server, but the Notifier connections is not updated.

If either the connection got suspended because of 15 seconds of
inactivity, or the connections gets closed while change_update_dn() is
running (too long), connect_to_ldap() called select_server() if the
original chosen server was not reachable.
On a slave that might select a different LDAP Backup server, but the
connection to the Notifier is *not* updated and still connects the old
host.
If the newly selected Backup is behind the previous host, the UDL will
see references to entries not yet replicated and will do arbitrary
things.

Drop the call to select_server().
Add error handling with 5 retries and return with an error instead of
looping infinitely.