Bug #47617 uvmm: Fix migration of snapshots
The libvirt API does not support migrating domains with snapshots.
Previously UCS carries a patch to work-around that by putting the
snapshot XML on a shared storage, only deleting the references on the
source host and adopting them on the destination.

That patch carries some bugs, which leads to the whole domain XML
getting lost during some migration scenarios, mainly offline migration.

Drop those patches and implement the following logic:
- backup the snapshot XML for disaster recovery.
- fetch the snapshot XML data into memory.
- delete only the snapshot meta data.
- migrate the domain, which now has no snapshots.
- Re-add the snapshot XML to the migrated domain.

Care must be taken as the snapshots have a hierarchy and build a tree:
- snapshots must be removed leaves to root.
- snapshots must be restored root to leaves.