Bug #48083: Disallow migration of multiple domains
The code to migrate multiple domains at once did not work. Only the first
one was migrated. Since it's most of the time not a good idea to migrate
multiple domains at the same time and we don't really have a queue for this
sort of think, it's now disallowed.