Bug #36661 uvmm: Fix parsing memory and disk capacities.
If '1.000.000.000 B' is entered, the old regular expression did not
match any unit multiplier [KMGTP] and returned null for group[2], so
"defaultUnit" was added, adding an additional "1M" multiplier for RAM.

This lead to "1.048.576.000.000.000" being written into the XML data,
overflowing an "int" an being returned as "0" on the next "get".

Only apply "defaultUnit" if no "B" or multiplier is specified.

Signed-off-by: Jürn Brodersen <brodersen@univention.de>