Bug #48340 uvmm: Normalize CPU utilization of host.
To calculate the host utilization the percentage of each domain must be
multiplied with the number of virtual CPUs per domain to get

To calculate the host utilization the following formula is used:

delta_t := length of this measurement interval
for each domain
  for each VCPU of that domain
    delta_used ?= number of CPU seconds of this measurement interval
    sum += delta_used
utilization := sum / delta_T / #CPU

UVMMd does neither track `delta_used` per domain nor is delta_t` the
same for all domains. But UVMMd tracks the utilization per domain:

    _cpu_usage ?= delta_used / delta_t / #vCPU

To get back `delta_used` we multiply with `#vCPU` and `delta_t := 1` for
simplicity, which leads to the following formula:

for each domain:
  sum += _cpu_usage * #vCPU
utilization := sum / #CPU

The old algorithm failed to take the number of vCPUs and host-CPUs into
account, which resulted into the utilization going over 100%.

Multiply the per-domain utilization with the number of vCPUs.
Normalize the host utilization by dividing by the number of host CPUs.
Remove the capping to 1000‰.